Fix None passed to LengthValidator
It is valid to pass None as a default value to the validator. It should
just return ValueError, which it will do with this change via the inner
validator.

Fixes https://github.com/galaxyproject/galaxy/issues/17961

Ideally we'd add type annotations, but with the current structure and
instantiation order it seems impossible to do this in a meaningful way.

If we refactored the validators to be functions, and validators
functions were registered per parameter type we could have narrow types on the
validator signature.