remove linter for unstripped text content for tool xml leaves
reverts #17656

we have to many xml nodes where this is allowed (e.g. command)
and limiting the linter to a statically defined list of node types
seemed difficult to maintain

replaced this with a change in the xsd that should lead to the
automatic removal of surplus white spaces (leading/trailing are removed,
multiple whitespaces are replaced by a single space) for some of the
node types that were the original motivation for the linter
(if this proves useful this could be used for more node types).