Improve test: enable db triggers; do not verify HistoryAudit table
We do not verify the counts of rows in the HistoryAudit table because
with the db trigger enabled, those rows are created automatically.

Although it is possible to calculate the expected counts, those numbers
would be dependend on previous db statements executed in the same test,
and that calculation would be not straightforward, making the test
brittle and confusing.