Add PersistentTaskProgressMonitorAlert component
The PersistentTaskProgressMonitorAlert component is added to provide a visual representation of the progress of a long-running operation. It includes different alert messages for different states of the task, such as in progress, completed, failed, and request failed. The component also supports automatic download of the task result when it is completed, if enabled.