Add expiration time to task monitors
The `expirationTime` specifies the time (in milliseconds) after which a task should be considered expired. Requests to check the task status after this time should be avoided as they are not guaranteed to return the correct status.

This is to avoid checking for tasks that were completed and then cleaned up from the tasks storage backend which will report the task as "pending" in the case of Celery tasks as they will be unknown after cleanup https://celery-safwan.readthedocs.io/en/latest/userguide/tasks.html?highlight=pending#pending