Introduce galaxy_external_url variable
Would fix
https://github.com/galaxyproject/galaxy/issues/18572#issuecomment-2239199496,
and prevent further abuse of galaxy_infrastructure_url, with the
downside that users would receive links that don't correspond to their
subdomain server if they had been using one.

The upside is that URLs can change, so maybe this is preferable over
recording the URL for jobs/invocations/histories etc, and it seems
messy to determine in what context one would like to see e.g.
a job finish message.