Set minimum weasyprint version
We're not pinning transitive conditional dependencies, and a new pydyf
version, which is a dependency of weasyprint, had a new release with a
breaking change, see (https://github.com/Kozea/WeasyPrint/issues/2200).

Newer releases of weasyprint have introduced max version pins, so I
think we're ok with just setting a minimum version here.