[24.1] Add error handling and initial loading state to Invocation State
Fixes https://github.com/galaxyproject/galaxy/issues/18723

Without this fix, if you get to the `WorkflowInvocationState` route for another user's invocation, there is a 404 error which is unhandled since there is a computed `invocation` ref that we constantly poll.
Adding a `initialLoading` ref that ensures the first fetch for the invocation was valid (or the error is caught) fixes this.