in fact do show the history in run form view; just say "Input History"
We removed the switch/view history button from the workflow run form version of the header, but it actually makes sense to keep it there since it lets the user know which history is going to be the input for the workflow (it changes as you change the history).

In the invocation version of course it always made sense because then it is scoped to the history that was used as input for that invocation.