fix bug for empty requirements
an empty set of requirements gives an empty requirement string which
was wrongly parsed to an CondaTarget with empty package name

This triggered a `conda search ''` which took so much memory that other
processes crashed.

The fix is to parse an empty list of CondaTargets for an empty target
string. In addition we check for empty package name during the creation
of the CondaTarget (to make this more easily detectable in the future)