Cleanup vega descriptions for workflow metrics.
It isn't exposed explictly but seeing this text when opening the Vega source or using the Vega editor from the link on the widget was a little offputting. I thought maybe the data was random at first but the jitter is only added in the x direction for reasons that make sense once you think about but isn't something I think should be forefronted.