Record implicitly converted dataset as input dataset
I believe this was always the intention and is how it works if the
dataset already exists (i.e. on a re-run). The converter records the
dataset that the user chose, so there's no gap in the provenance either.

This somewhat addresses
https://github.com/galaxyproject/total-perspective-vortex/issues/141
so you can (reliably) differetiate your rules on the input datatype
and filesize combination.