Fix target_type in auto-conversion context
which means we collect and store the converted dataset as the job input.

We don't need to wait for the galaxy.json collection, we know the exact
target type already. That fixes the retrieval in
`get_converted_files_by_type`.

I believe this was always the intention and is how it works if the
dataset already exists (i.e. on a re-run), and for all converters that
don't use galaxy.json The converter records the dataset that the user
chose, so there's no gap in the provenance either.

This somewhat addresses
https://github.com/galaxyproject/total-perspective-vortex/issues/141
so you can (reliably) differetiate your rules on the input datatype
and filesize combination.