Don't show published workflows in workflow activity of workflow editor
If we want to do this we need to separate owned and published, then make
a new copy or reference as we insert the public workflow. As it
stands you won't be finding your own workflows, since the public
workflows are far more numerous then any one user's workflows.

Fixes https://github.com/galaxyproject/galaxy/issues/19265