mulled: also consider non strict channel priority
for the creation of conda environments we try

1. strict channel priotity
2. if this fails try without

https://github.com/galaxyproject/galaxy/blob/5f484fca6255fbbc8a66016601c3fcab20271c81/lib/galaxy/tool_util/deps/conda_util.py#L282

For the creation of mulled containers we don't do this.
I stumbled over this a few times when creating containers
for older tools, where a bioconda package was used that
is now in conda-forge (and with strict priority only the
newer conda-forge packages can be installed)

Also restructured slightly. Before the `build` was appended
to the `involucro_args` early. Therefore some inserts
were needed in the middle of the list.

Fix verbose option which was ignored before.