extend repeat test for min=0 and default=1
it should be possible to execute tool without
any repeat elements.

but tool tests require a value for the data parameter
even if no repeat is given.

I guess this broke in https://github.com/galaxyproject/galaxy/pull/17018
But if I revert the backend change tools do not
load anymore.