Fix version flicker on job rerun
Fixes https://github.com/galaxyproject/galaxy/issues/19477,
which should be a problem for built-in tools where the rerun shouldn't
use the latest version. This is because built-in tools don't contain the
version in the tool id, and we didnt' persist the verson in the ToolForm
data. Made worse by https://github.com/galaxyproject/galaxy/issues/18848#issuecomment-2393830406