Fix WSGI response status handling in controller methods
Fixes https://github.com/galaxyproject/galaxy/issues/19519.
If a string is being set and can't just be the status code cast to a
string (which would be correct for ASGI), as that fails in a2wsgi.

Easiest and most consistent is to just set the status code to an int and
have the wsgi_status method look up the correct description.