Revert "Split installation shell command into four"
This reverts commit bd24d0cafc14ba347e9a50713ebece6b630e7dc1.

It makes sense to revert the changes because every `run()` statement runs in its own container, thus it cannot be ruled out that other code is building the preinstall or postinstall commands under the assumption that they will run in the same container (or even in the same container as the conda commands). An example is available on https://github.com/galaxyproject/galaxy/pull/19545#discussion_r1944557524.