Make tool ids readable in invocation metrics
The whole tool id is still viewable on hover, in the job details and in
the vega editor.

Also fixes sorting by step index and replaces the pie chart with stacked
bar chart. Pie charts get a bad name in statistics and datascience.
In this case I think the same purpose can be fulfilled with a stacked
barchart, which allows us to also represent the cumulative time on the
y-axis. The bar chart categories are sorted by cumulative y value.