Improved simplicity and isolation in transiently failing test.
The test failure without any really useful context on the client is bad and indicating a problem - it is going to be some concurrency or database thing in the test environment or an issue with us hitting Github a bunch right in a row. Regardless - it isn't the job of this test case to uncover that I think - I think isolating the network and simplifying the test here will increase the test robustness and ensure it is testing what it is meant to test.