Fix conditional dependency path resolution
The path resolution logic was not correct. A relative path within
galaxy.yml is relative to the config directory. To fix this we now rely
on GalaxyAppConfiguration which does all of the path resolution exactly
as Galaxy would.

I noticed this when I added a file source and the conditional depdency
logic would not attempt to install the correctly defined conditional
dependency. I did have `file_sources_config_file` set in my galaxy
config. This would have only worked if I left that setting undefined and
we correctly joined the path with the config dir or if I had set an
absolute path.