Remove ``num_unique_values`` tiff metadata element
This was added in https://github.com/galaxyproject/galaxy/pull/18951 but
it's memory consumption scales with input size. On a 5.5GB TIFF file
this consumed a peak of 18GB of memory.

@kostrykin added a workaround for the memory consumption in
https://github.com/galaxyproject/galaxy/pull/19830, but this now very
slow.

Besides improving the efficient, Perhaps there is a more gentle way to approach

> Many tools assume that images are binary, or label maps, for example, and it makes no sense to run those tools on other images.

?

Could those tools just error out, and/or consume a different datatype ?
Do we need the full unique values, or is it ok to just check if it's
more than `n` unique values ?