Remove ``num_unique_values`` tiff metadata element
This was added in https://github.com/galaxyproject/galaxy/pull/18951 but
it's memory consumption scales with input size. On a 5.5GB TIFF file
this consumed a peak of 18GB of memory.

A working and more memory efficient implentation would be
```
import rasterio
import numpy as np
from collections import Counter

def count_unique_values_rasterio(file_path):
    counter = Counter()

    with rasterio.open(file_path) as src:
        for _, window in src.block_windows(1):  # Read in small windows
            data = src.read(1, window=window)  # Read only the small chunk
            unique, counts = np.unique(data, return_counts=True)
            counter.update(dict(zip(unique, counts)))

    return len(counter)
```

However this seems to be a property mostly useful for test assertions
(do correct me if I'm wrong @kostrykin), as such I would not compute
this for every tiff dataset, and instead handle it in the client test
code.
Considering that the rasterio wheel is 22MB in size and I haven't found
any use in the current test framework I think it's better to remove this
for now and add it when necessary.