Revise consistently failing edam tool panel view test.
The tool panel view renders the default view before switching to the EDAM I think, so retrying the assertion instead of just retrying any DOM click failures should allow that to stabilize. The UI flicker seems to be a regression though - but this test didn't prevent that so now it has to test something else.

Also I think the EDAM stuff is just on by default now so this doesn't need to be an integration test - as such moved it to the vanilla Selenium tests.