Guard state update with limit queries
These are essentially the same queries that are done in
`JobHandler.__check_user_jobs`, `JobHandler.__check_destination_jobs`
etc, but now it's all in in a single update statement.

I suppose performance might be a concern, however we still run through
the (cached) checks before we decide to queue the job, so I think
the cost is likely minimal. By integrating the limit check in the query
i think it should become very unlikely that jobs can bypass limits in a
multi handler scenario.