Don't fail startup on conflicting data table definition
While I'm all for failing hard on configuration issues, this one is
almost unpredictable, as rebooting a galaxy instance will fail if being
unlucky enough to install the wrong combination of tools with data
tables. And it won't fail during installation, but at any point a
restart is necessary, making it not obvious what caused this and when to
expect this.