Implement galaxy-job-config-init.
This is a script I want to use in Planemo to rapidly build a functional, production job_conf.yml from command-line arguments. I want to get it as robust as I can without tweaks so it can be used directly from "planemo run workflow.ga" in many scenarios for production workflow runs.

That is optimistic, so I will also implement a "planemo job_config_init" command to create a file for a workflow project or "Planemo profile" that can be edited by the user and picked up in subsequent workflow runs. But for ``pip install galaxy``, I've included ``galaxy-job-config-init`` as a script here that can be used admins to rapidly prototype job configs using the same code.

It is intractable to test all of these job configurations - so I've implemented the ability to quickly review a bunch of expected common invocations. I've included the resulting Markdown below. I haven't configured a Galaxy to talk to a cluster since before containers existed and before the job conf had even an XML configuration - so I'd love a review and advice from actual admins.