invocation share: only make history and workflow shareable if not already
Note: We will mostly have a summarized history object without the `importable` property and in that case we still attempt to make it shareable. Only case I can see right now where the history will have the `importable` property is when the history is current (and hence, we have a detailed object).