Migrate `JobFilesAPIController` to FastAPI
`FastAPIJobFiles` is the new, FastAPI version of `JobFilesAPIController`. It should exhibit exactly the same behavior as `FastAPIJobFiles`.

Something to watch out for is that FastAPI has some extra built-in features that the legacy WSGI system did not have, such as answering `HEAD` requests or returning HTTP status code 405 when unsupported request methods are provided (e.g. `PROPFIND`, used for WebDAV). However, those do not work because of the way legacy WSGI endpoints are injected into the FastAPI app (using `app.mount("/", wsgi_handler)`), meaning that such requests are passed to the `wsgi_handler` sub-application.