Improve compliance of `JobFilesAPIController` with the HTTP spec (`files` endpoint)
This commit introduces the following differences in behavior:
- `/api/jobs/{job_id}/files` now answers, in addition to `GET` requests as it used to, also `HEAD` requests (read motivation for this change below).
- `/api/jobs/{job_id}/files` now answers `PROPFIND` requests with HTTP status code 501 (read motivation for this change below).

The reason behind the code explicitly supporting `HEAD` requests and answering `PROPFIND` requests with status code 501 is an unfortunate interaction between the ARC remote job runner that is under development, the behavior of legacy API endpoints and how they are integrated within the FastAPI app.

The ARC remote job runner (which will be implemented as `lib.galaxy.jobs.runners.pulsar.PulsarARCJobRunner`) expects this endpoint not only to answer `GET` and `HEAD` requests, but also even to return HTTP codes other than 404 when `PROPFIND` requests are issued. The latter are not part of the HTTP spec, but it is used in the WebDAV protocol. The correct answer to such requests is likely 501 (not implemented).

FastAPI answers HEAD requests automatically for `GET` endpoints, and returns HTTP 405 (method not allowed) for `PROPFIND`, which maybe is not fully correct but tolerable because it is one less quirk to maintain.

However, because of the way legacy WSGI endpoints are injected into the FastAPI app (using `app.mount("/", wsgi_handler)`), the built-in support for `HEAD` requests and for returning HTTP 405 for `PROPFIND` breaks, because such requests are passed to the `wsgi_handler` sub-application. This means that the endpoint still needs to include some code to handle this behavior.

When ALL routes have been migrated to ASGI (no WSGI handler sub-application needed anymore), some lines of code can be removed, they are labeled using comments.