close the readme editor when switching to report editor
Maybe do this for other activities as well? E.g. inputs, tools...

Of course, users can simply close the readme view with the prominent exit button, but there are case when it could get confusing if it's in the way: e.g.: Editing/Viewing readme, user changes to tools activity, clicks on a tool and though the tool is added to the graph, they do not notice since the readme is in the way.