Pin limits to >=2.5.0
`limits` is a transitive dependency (from slowapi) that currently
depends on `packaging>=21,<25` .
Now that `packaging` 25.0 was released, uv prefers to upgrade `packaging`
and downgrade `limits` to 2.4.0 (which didn't depend on `packaging`).
This will be fixed by https://github.com/alisaifee/limits/pull/274 ,
but pinning `limits` to >=2.5.0 will prevent similar issues in the future
as well.