match types generated by pydantic model = no more type aliases for client
As recommended, matched pydantic types. What I mean by "no more type aliases for client" is that if you see the changes in this commit, before this commit, we were allowing several aliases in the workflow landing data `request_state` to be shown in the client components (e.g.: `ext`, `filetype` and `extension` were all valid keys to show the file extension, any of `url` or `location` would work for the source URL, etc).

Now, we have stricter types hence, the workflow landing (for the client workflow run form) requires exactly the expected structure in `src/components/Form/Elements/FormData/types.ts`.

If any landing data differs from this, it is not shown.

Co-authored-by: mvdbeek <m.vandenbeek@gmail.com>